requirements.txt
streamlit #==1.44.1
streamlit-js-eval #==0.1.7
#crewai==0.193.2
#crewai_tools==0.76.0
#openai==1.90.0
#pynvml==12.0.0
#easygui==0.98.3
#pypandoc_binary==1.15
#ruamel_yaml==0.18.14
#keybert==0.9.0
#hf_transfer==0.1.9
#deepeval==3.6.8
#evaluate==0.4.6
#rouge_score==0.1.2
#bert_score==0.3.13
#codebleu==0.7.0
#qdrant-client==1.15.1
#dataframe_image==0.2.7

crewai==0.193.2
crewai_tools==0.76.0
openai==1.90.0
pynvml==12.0.0
easygui==0.98.3 #==0.98.3
pypandoc_binary==1.15
ruamel_yaml #==0.18.14
keybert==0.9.0
hf_transfer==0.1.9
deepeval==3.6.8
evaluate==0.4.6
rouge_score==0.1.2
bert_score==0.3.13
codebleu==0.7.0
qdrant-client==1.15.1
dataframe_image==0.2.7
chromadb==1.0.13
# chromadb-pysqlite3 #==0.5.3
# pysqlite3-binary #==0.5.4
jsonpath_ng #==1.7.0
PyGithub==2.8.1
more_itertools==10.8.0
pytesseract==0.3.13
jsonpath_ng==1.7.0
validators==0.35.0